Add @META{...} and META{...}[\d] interpolation for arrayrefs
This was prompted by the Dist::Zilla 4.300039 release changing
how the license is included in distmeta. That change caused
Dist::Zilla to conform to the CPAN::Meta spec, and our tests
to fail.

The license is an arrayref now, so we need some way to access
array values in distmeta. There are now two.

First, you can now use [\d] to access an arrayref element, just
like you could use {...} to access a hashref element. Simply
use META{license}[0] to access the zeroth element in the license
arrayref.

Second, you can use the @META modifier. Like the !META modifier,
this does something special to the value you extract. If you
extract an arrayref, you may add the @ before META to tell us
to stringify it. So, @META{license} means to substitute in the
value "@{ $distmeta->{license} }".

The documentation was updated to provide examples, and a warning
about taking responsibility for accessing the distmeta structure
properly, with a pointer to relevant CPAN::Meta documentation.

The tests were updated to use both of the new arrayref META
substitution methods.